fix: improve vmlens handling (#1628)
* fix: for #1611 Only execute concurrent tests with VMLens

VMLens test run now takes 4s and there is no risk of hanging tests.

The hanging tests was during executing ArchUnitTests with VMLens.

Renamed  VmLensTest to VmLensCT and added a filter in the VMlens plugin.

increased timeout for gherkin tests to 30s from 10s

Signed-off-by: ThomasKrieger <thomas.krieger@vmlens.com>

* added await in ProviderSteps
and removed Awaitility.await() since it is
not needed anymore since we wait during the setting of the states

Signed-off-by: ThomasKrieger <thomas.krieger@vmlens.com>

---------

Signed-off-by: ThomasKrieger <thomas.krieger@vmlens.com>